requirements.txt
tensorboard==2.14.0
numpy==1.22.4
pyglet==1.3.2
# gym==0.15.3 # for FetchPush-v1
gym==0.8.2 # for Original Env
mpi4py==3.1.3
ipdb
transforms3d
omegaconf
opencv-python
attrdict
natsort
lxml
hydra-core==1.2.0
mpl-tools==0.2.57

# ------ for Pendulum ----
# gym==0.8.2 #
